Backend Api Completed

+++ controllers/contactController.js
@@ -1,18 +1,11 @@
 const asyncHandler = require("express-async-handler");
 const Contact = require("../models/contactModels")
 const getContact = asyncHandler(async(req, res) => {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({ user_id: req.user.id });
     console.log(contacts);
     res.status(200).json(contacts);
 })
-const getContactById = asyncHandler(async(req, res) => {
-    const contact = await Contact.findById(req.params.id)
-    if (!contact) {
-        res.status(404);
-        throw new Error("Contact not found")
-    }
-    res.status(200).json(contact)
-})
+
 
 const createContact = asyncHandler(async(req, res) => {
     console.log(req.body);
@@ -25,15 +18,30 @@ const createContact = asyncHandler(async(req, res) => {
         name,
         email,
         phone,
+        user_id: req.user.id,
     });
     res.status(200).json(contact);
 })
+
+const getContactById = asyncHandler(async(req, res) => {
+    const contact = await Contact.findById(req.params.id)
+    if (!contact) {
+        res.status(404);
+        throw new Error("Contact not found")
+    }
+    res.status(200).json(contact)
+})
+
 
 const updateContact = asyncHandler(async(req, res) => {
     const contact = await Contact.findById(req.params.id)
     if (!contact) {
         res.status(404);
         throw new Error("Contact not found")
+    }
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User dont have permission")
     }
     const updatedcontact = await Contact.findByIdAndUpdate(
         req.params.id,
@@ -47,6 +55,10 @@ const deleteContact = asyncHandler(async(req, res) => {
     if (!contact) {
         res.status(404);
         throw new Error("Contact not found")
+    }
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User dont have permission")
     }
     await Contact.deleteOne({ _id: req.params.id });
     res.status(200).json(contact);
